requirements.txt
discord-qalib==2.5.0
discord.py>=2.2.3
python-dotenv~=1.0.0
sqlalchemy~=2.0.13
Pint~=0.21
babel~=2.12.1
Jinja2~=3.1.2
asyncio~=3.4.3
pydantic>=2.1.0
